test/menu_extractor/requirements.txt
langchain-core>=0.1.16
langchain-anthropic>=0.1.1
langchain-openai>=0.1.1
langchain>=0.1.5
langgraph>=0.0.27
pydantic>=2.0.0
beautifulsoup4>=4.12.2
playwright>=1.40.0
python-dotenv>=1.0.0 